move classes to own cpp and hpp file, fix formatting of the files(clang-format doesn't seem to work correctly)
